requirements.txt
fastapi
uvicorn[standard]
langchain-community
langchain-text-splitters
langchain-ollama
qdrant-client
pymongo
openai
pypdf
python-dotenv